mid-point circle drawing algorithm
The mid-point circle drawing algorithm is an algorithm used to determine the points needed for rasterizing a circle. 
 
We use the mid-point algorithm to calculate all the perimeter points of the circle in the first octant and then print them along with their mirror points in the other octants. This will work because a circle is symmetric about it’s centre.